New docker compose file
