+++ compose.yaml
@@ -1,40 +1,17 @@
-# Compose file for Verifiable Credentials demos
 name: vcdemo
 services:
-  caddy:
-    image: caddy:latest
+  vcdemo:
+    image: gcr.io/distroless/static-debian12
+    container_name: vcdemo
+    user: 1000:1000
     restart: unless-stopped
-    container_name: "caddy"
-
-    # Expose both ports so automatic certificate provisioning works
-    # But the only one working is the 443, as requests to 80 will be redirected to the 443
-    ports:
-      - "80:80"
-      - "443:443"
-    volumes:
-      - ./Caddyfile:/etc/caddy/Caddyfile
-      - caddy_data:/data
-      - caddy_config:/config
-
-  vcbackend:
-    build: .
-    image: vcdemo:latest
-    restart: unless-stopped
-    container_name: "vcbackend"
     ports:
       - "3030:3030"
       - "8090:8090"
       - "9999:9999"
       - "9998:9998"
     volumes:
-      - ./data:/app/data
+      - .:/dome
+    working_dir: /dome
+    entrypoint: ["/dome/vcdemo", "serve"]
 
-# Caddy requires write access to two locations: a data directory, and a configuration directory.
-# While it's not necessary to persist the files stored in the configuration directory, it can be convenient.
-# However, it's very important to persist the data directory.
-# Caddy stores there TLS certificates, private keys, OCSP staples, and other necessary information.
-# Defining the data volume as external makes sure 'docker compose down' does not delete the volume.
-# You may need to create it manually using docker 'volume create caddy_data'.
-volumes:
-  caddy_data:
-  caddy_config:
