+++ www/index.html
@@ -45,7 +45,7 @@
   "MicroWallet": "/pages/MicroWallet-F4GFBQ25.js",
   "SIOPSelectCredential": "/pages/SIOPSelectCredential-WLBAVYBM.js",
   "SWNotify": "/pages/SWNotify-GLIQS6YO.js",
-  "ScanQrPage": "/pages/ScanQrPage-3PNFDNWM.js",
+  "ScanQrPage": "/pages/ScanQrPage-FWSM7CP2.js",
   "SelectCamera": "/pages/SelectCamera-PXJHLD5U.js",
   "TestPage": "/pages/TestPage-OKISILJI.js"
 }
